Added some bussiness logic

+++ src/main/java/com/example/testtasktemvoy/service/impl/OrderServiceImpl.java
@@ -9,12 +9,15 @@ import com.example.testtasktemvoy.model.Order;
 import com.example.testtasktemvoy.model.OrderProduct;
 import com.example.testtasktemvoy.model.OrderStatus;
 import com.example.testtasktemvoy.model.Product;
+import com.example.testtasktemvoy.model.Role;
 import com.example.testtasktemvoy.model.User;
 import com.example.testtasktemvoy.repository.OrderProductRepository;
 import com.example.testtasktemvoy.repository.OrderRepository;
 import com.example.testtasktemvoy.repository.ProductRepository;
 import com.example.testtasktemvoy.service.OrderService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import javax.persistence.EntityNotFoundException;
@@ -22,6 +25,9 @@ import javax.transaction.Transactional;
 import java.util.Comparator;
 import java.util.InputMismatchException;
 import java.util.List;
+import java.util.Objects;
+import java.util.Timer;
+import java.util.TimerTask;
 import java.util.stream.Collectors;
 
 @Service
@@ -30,8 +36,8 @@ import java.util.stream.Collectors;
 public class OrderServiceImpl implements OrderService {
 
     private final OrderRepository orderRepository;
-    private final OrderProductRepository orderProductRepository;
     private final ProductRepository productRepository;
+    private final Timer timer = new Timer();
 
     @Override
     public void createOrder(CreateOrderDto createOrderDto, User user) {
@@ -40,25 +46,43 @@ public class OrderServiceImpl implements OrderService {
         order.setComment(createOrderDto.getComment());
         order.setUser(user);
 
-        List<OrderProduct> orderProductList = getOrderProductsFromDto(createOrderDto.getCreateOrderProductDtoList(), order);
+        List<OrderProduct> orderProductList = getOrderProductsFromDto(createOrderDto.getProducts(), order);
 
         order.setOrderProducts(orderProductList);
-        orderRepository.save(order);
-        orderProductRepository.saveAll(orderProductList);
+        Order savedOrder = orderRepository.save(order);
+
+        deleteAfterTenMinutesIfUnpaid(savedOrder.getId());
+    }
+
+    private void deleteAfterTenMinutesIfUnpaid(Long id){
+        TimerTask task = new TimerTask() {
+            @Override
+            public void run() {
+                Order order = orderRepository.findById(id).orElseThrow(() -> new EntityNotFoundException(ErrorMessage.ORDER_DOESNT_EXIST + id));
+                if(order.getOrderStatus().equals(OrderStatus.UNPAID)){
+                    restoreProductsQuantity(order.getOrderProducts());
+                    orderRepository.delete(order);
+                }
+            }
+        };
+
+        long delay = 1000 * 60 * 10;
+        timer.schedule(task, delay);
     }
 
     @Override
     public void updateOrder(Long id, CreateOrderDto updateOrderDto) {
         Order order = orderRepository.findById(id).orElseThrow(() -> new EntityNotFoundException(ErrorMessage.ORDER_DOESNT_EXIST + id));
+
+        checkIfPrincipalHasAccess(order);
+
         order.setComment(updateOrderDto.getComment());
 
         restoreProductsQuantity(order.getOrderProducts());
-        orderProductRepository.deleteAll(order.getOrderProducts());
 
-        List<OrderProduct> orderProductList = getOrderProductsFromDto(updateOrderDto.getCreateOrderProductDtoList(), order);
+        List<OrderProduct> orderProductList = getOrderProductsFromDto(updateOrderDto.getProducts(), order);
 
         order.setOrderProducts(orderProductList);
-        orderProductRepository.saveAll(orderProductList);
         orderRepository.save(order);
     }
 
@@ -81,8 +105,10 @@ public class OrderServiceImpl implements OrderService {
     @Override
     public void deleteOrder(Long id) {
         Order order = orderRepository.findById(id).orElseThrow(() -> new EntityNotFoundException(ErrorMessage.ORDER_DOESNT_EXIST + id));
+
+        checkIfPrincipalHasAccess(order);
+
         restoreProductsQuantity(order.getOrderProducts());
-        orderProductRepository.deleteAll(order.getOrderProducts());
         orderRepository.delete(order);
     }
 
@@ -90,6 +116,15 @@ public class OrderServiceImpl implements OrderService {
     public void markOrderAsPaid(Long id) {
         if(!orderRepository.existsById(id)) throw new EntityNotFoundException(ErrorMessage.ORDER_DOESNT_EXIST + id);
         orderRepository.setOrderStatus(id, OrderStatus.PAID.toString());
+    }
+
+    private void checkIfPrincipalHasAccess(Order order){
+        User currentUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if(!Objects.equals(currentUser.getId(), order.getUser().getId())
+                && !currentUser.getRole().equals(Role.ADMIN)){
+            throw new AccessDeniedException("You can update only your own orders");
+        }
     }
 
     private List<OrderProduct> getOrderProductsFromDto(List<CreateOrderProductDto> createOrderProductDtoList, Order order){
@@ -113,6 +148,7 @@ public class OrderServiceImpl implements OrderService {
                 o -> {
                     Integer currentQuantity = o.getPk().getProduct().getItemsLeft();
                     o.getPk().getProduct().setItemsLeft(currentQuantity + o.getQuantity());
+                    productRepository.save(o.getPk().getProduct());
                 }
         );
     }
